Replace eslint with automatic-semicolon-insertion.
eslint was blocking our attempt to build a browser version since it
uses `fs.readdirSync`, which browserify doesn’t polyfill. The
automatic-semicolons-insertion library is also much smaller.
