+++ package.json
@@ -31,11 +31,12 @@
   "license": "MIT",
   "dependencies": {
     "add-variable-declarations": "^1.3.1",
-    "babel-eslint": "^5.0.0",
+    "ast-processor-babylon-config": "^1.0.0",
+    "automatic-semicolon-insertion": "^1.0.1",
+    "babylon": "^6.7.0",
     "coffee-lex": "^1.3.10",
     "decaffeinate-parser": "^1.2.13",
     "detect-indent": "^4.0.0",
-    "eslint": "^1.10.3",
     "esnext": "^1.17.2",
     "magic-string": "^0.10.2",
     "repeating": "^2.0.0"
