+++ src/index.js
@@ -1,5 +1,5 @@
 import AddVariableDeclarationsStage from './stages/add-variable-declarations/index.js';
-import EslintStage from './stages/eslint/index.js';
+import SemicolonsStage from './stages/semicolons/index.js';
 import EsnextStage from './stages/esnext/index.js';
 import MainStage from './stages/main/index.js';
 import NormalizeStage from './stages/normalize/index.js';
@@ -29,7 +29,7 @@ export function convert(source: string, options: ?Options={}): ConversionResult 
     NormalizeStage,
     MainStage,
     AddVariableDeclarationsStage,
-    EslintStage,
+    SemicolonsStage,
     EsnextStage
   ]);
 }
